+++ tests/test_06_math/test_630_bezier4p.py
@@ -2,11 +2,12 @@
 # License: MIT License
 import pytest
 import math
-from ezdxf.math import ConstructionEllipse, Matrix44, Vector, Vec2
-from ezdxf.math.bezier4p import (
-    Bezier4P, cubic_bezier_arc_parameters, cubic_bezier_interpolation,
-    cubic_bezier_from_arc, cubic_bezier_from_ellipse
+from ezdxf.math import (
+    ConstructionEllipse, Matrix44, Vec3, Vec2,
+    Bezier4P, cubic_bezier_interpolation, cubic_bezier_from_arc,
+    cubic_bezier_from_ellipse,
 )
+from ezdxf.math.bezier4p import cubic_bezier_arc_parameters
 
 DEFPOINTS2D = [(0., 0.), (3., 0.), (7., 10.), (10., 10.)]
 DEFPOINTS3D = [(0.0, 0.0, 0.0), (10., 20., 20.), (30., 10., 25.),
@@ -149,7 +150,7 @@ def test_reverse():
 def test_transform_interface():
     curve = Bezier4P(DEFPOINTS3D)
     new = curve.transform(Matrix44.translate(1, 2, 3))
-    assert new.control_points[0] == Vector(DEFPOINTS3D[0]) + (1, 2, 3)
+    assert new.control_points[0] == Vec3(DEFPOINTS3D[0]) + (1, 2, 3)
     assert new.control_points[0] != curve.control_points[
         0], 'expected a new object'